// sealos/frontend/providers/dbprovider/src/types/db.d.ts
export type DBType = `${DBTypeEnum}`;

export type SupportMigrationDBType = Extract<DBType, 'postgresql' | 'mongodb' | 'apecloud-mysql'>;

export type SupportConnectDBType = Extract<DBType, 'postgresql' | 'mongodb' | 'apecloud-mysql'>;

export type SupportReconfigureDBType = Extract<
  DBType,
  'postgresql' | 'mongodb' | 'apecloud-mysql' | 'redis'
>;

export type DeployKindsType =
  | V1Deployment
  | V1StatefulSet
  | V1ConfigMap
  | V1Service
  | V1Ingress
  | V1Secret
  | V1HorizontalPodAutoscaler;

export type EditType = 'form' | 'yaml';

export interface DBStatusMapType {
  label: string;
  value: `${DBStatusEnum}`;
  color: string;
  backgroundColor: string;
  dotColor: string;
}
// ...
export interface DBEditType {
  dbType: DBType;
  dbVersion: string;
  dbName: string;
  replicas: number;
  cpu: number;
  memory: number;
  storage: number;
  labels: { [key: string]: string };
  terminationPolicy: KubeBlockClusterTerminationPolicy;
  autoBackup?: AutoBackupFormType;
  dataSourceId?: number;
}

export type DBSourceType = 'app_store' | 'sealaf';

export type DBSource = {
  hasSource: boolean;
  sourceName: string;
  sourceType: DBSourceType;
};

export interface DBDetailType extends DBEditType {
  id: string;
  createTime: string;
  status: DBStatusMapType;
  conditions: DBConditionItemType[];
  isDiskSpaceOverflow: boolean;
  labels: { [key: string]: string };
  source: DBSource;
  totalCpu: number;
  totalMemory: number;
  totalStorage: number;
}

export interface DBConditionItemType {
  lastTransitionTime: string;
  message: string;
  observedGeneration: 3;
  reason: string;
  status: 'True' | 'False';
  type: string;
}